i remove all #

+++ part4/app/api/v1/reviews.py
@@ -26,8 +26,8 @@ class ReviewList(Resource):
             return {'error': 'Place not found'}, 404
         review = facade.get_all_reviews()
         try:
-            #if current_user == place.owner.id:
-                #return {'error': "you cannot review your own place."}, 400
+            if current_user == place.owner.id:
+                return {'error': "you cannot review your own place."}, 400
             if review:
                 return {'error': "You have already reviewed this place."}, 400
             new_review = facade.create_review(review_data)
@@ -57,8 +57,8 @@ class ReviewResource(Resource):
         review_data = api.payload
         if not review_data:
             return {'error': 'Review not found'}, 404
-        #if review.user.id != user_id:
-                #return {'error': "Unauthorized action."}, 403
+        if review.user.id != user_id:
+                return {'error': "Unauthorized action."}, 403
         try:
             updated_review = facade.update_review(review_id, review_data)
             return {'message': 'Review updated successfully'}, 200
@@ -68,13 +68,13 @@ class ReviewResource(Resource):
     @jwt_required()
     def delete(self, review_id):
         current_user = get_jwt_identity()
-        #user_id = current_user['id']
+        user_id = current_user['id']
         review = facade.get_review(review_id)
         review_data = api.payload
         if not review:
             return {'error': 'Review not found'}, 404
-        #if review.user.id != user_id:
-            #return {'error': "Unauthorized action."}, 403
+        if review.user.id != user_id:
+            return {'error': "Unauthorized action."}, 403
         deleted = facade.delete_review(review_id)
         return {'message': 'Review deleted successfully'}, 200
